Fix round(double precision) in both timing queries [skip render]
percentile_cont returns double precision and round(x, n) has no
double-precision form. The cast goes on the percentile, not the ORDER
BY. Caught by psql exit 3 on research-sql run 172.

+++ research/bettor_followup_timing.sql
@@ -85,10 +85,10 @@ UNION ALL
 
 SELECT 'F_DELAY_QUANTILES',
        m.horizon_s::text || 's',
-       'p50=' || round(percentile_cont(0.5) WITHIN GROUP (
-                  ORDER BY m.actual_lag_s::numeric), 1)::text
-       || ' p90=' || round(percentile_cont(0.9) WITHIN GROUP (
-                  ORDER BY m.actual_lag_s::numeric), 1)::text
+       'p50=' || round((percentile_cont(0.5) WITHIN GROUP (
+                  ORDER BY m.actual_lag_s::numeric))::numeric, 1)::text
+       || ' p90=' || round((percentile_cont(0.9) WITHIN GROUP (
+                  ORDER BY m.actual_lag_s::numeric))::numeric, 1)::text
        || ' max=' || round(max(m.actual_lag_s::numeric), 1)::text
        || ' n=' || count(*)::text
   FROM bettor_state_mids m
